+++ src/codeLists/nodeTypeCodeList.ts
@@ -1,8 +1,8 @@
-const nodeTypesCodeList = {
+const nodeTypeCodeList = {
     ['P']: 'Pysäkki',
     ['X']: 'Risteys',
     ['-']: 'Kuntaraja',
     ['E']: 'Pois käytöstä',
 };
 
-export default nodeTypesCodeList;
+export default nodeTypeCodeList;
