+++ src/components/sidebar/networkView/linkView/LinkView.tsx
@@ -10,7 +10,7 @@ import LinkService from '~/services/linkService';
 import NodeType from '~/enums/nodeType';
 import SubSites from '~/routing/subSites';
 import routeBuilder from '~/routing/routeBuilder';
-import municipalityCodeList from '~/codeLists/municipalitiesCodeList';
+import municipalityCodeList from '~/codeLists/municipalityCodeList';
 import navigator from '~/routing/navigator';
 import { LinkStore } from '~/stores/linkStore';
 import { MapStore } from '~/stores/mapStore';
