+++ src/components/sidebar/networkView/nodeView/NetworkNode.tsx
@@ -9,7 +9,7 @@ import LinkService from '~/services/linkService';
 import { Button, Dropdown } from '~/components/controls';
 import NodeType from '~/enums/nodeType';
 import NodeService from '~/services/nodeService';
-import nodeTypesCodeList from '~/codeLists/nodeTypesCodeList';
+import nodeTypeCodeList from '~/codeLists/nodeTypeCodeList';
 import NodeLocationType from '~/types/NodeLocationType';
 import ButtonType from '~/enums/buttonType';
 import Loader from '~/components/shared/loader/Loader';
@@ -119,7 +119,7 @@ class NetworkNode extends React.Component<INetworkNodeProps, InetworkNodeState> 
                                     onChange={this.onChange('type')}
                                     disabled={isEditingDisabled}
                                     selected={node.type}
-                                    codeList={nodeTypesCodeList}
+                                    codeList={nodeTypeCodeList}
                                 />
                             </div>
                         </div>
